refactor: make separate package for used models
