+++ src/models/event.py
@@ -1,0 +1,24 @@
+import json
+from typing import Dict
+from .timeline_component import TimelineComponent
+from .date import Date
+
+class Event(TimelineComponent):
+    def __init__(self, id: str, title: str, date: Dict, importance: str = "MEDIUM"):
+        super().__init__(id, title, importance)
+        self.date = Date(date)
+
+    def to_dict(self) -> dict:
+        base_dict = super().to_dict()
+        base_dict["date"] = {
+            "year": self.date.year,
+            "month": self.date.month,
+            "day": self.date.day
+        }
+        return base_dict
+
+    def export_json(self, filename: str = None):
+        if filename is None:
+            filename = f"{self.id}.json"
+        with open(filename, 'w') as f:
+            json.dump(self.to_dict(), f, indent=2)
